requirements.txt
astroid==2.3.3
Django==2.2.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
gunicorn==20.0.2
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mutagen==1.42.0
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.18
Pillow==6.2.1
protobuf==3.10.0
pycryptodome==3.9.1
pylint==2.4.4
PyMySQL==0.9.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wrapt==1.11.2
